+++ frontend/src/components/admin/AddNewService.js
@@ -4,7 +4,7 @@ import axios from 'axios'
 import swal from 'sweetalert';
 import './service.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
-import { Card,Button,Container,Form,CardGroup,Row,Image} from 'react-bootstrap';
+import { Card,Button,Modal,Form,CardGroup,Row,Image} from 'react-bootstrap';
 import Loading from '../../components/Loading';
 
 
@@ -20,6 +20,7 @@ export default function AddNewService () {
     const [name,setName] = useState();
     const [IMage,setImage]=useState();
     const [Description,setDescription] = useState();
+    const [show,setShow] =useState(false);
 
     // i want to display all information in screen 
     useEffect(()=>{ 
@@ -53,19 +54,38 @@ const handelAddNewService =()=>{
 // delete service
 
     const handelDeleteService =(id)=>{
+
+     
+// 
+swal({
+  title: "Are you sure?",
+  text: "Once deleted, you will not be able to recover this imaginary file!",
+  icon: "warning",
+  buttons: true,
+  dangerMode: true,
+})
+.then((willDelete) => {
+  if (willDelete) {
     axios.delete(`http://localhost:3030/services/${id}/delete`)
-    .then((res)=>{
-        setNewService(res.data)
-        swal({
-            title:'Service is deleted ',
-            icon:'success'
-          })
-    })
+.then((res)=>{
+    swal(" Your file has been deleted!", {
+    icon: "success",
+
+})
+setNewService(res.data)
+    });
+  } else {
+    swal("Your file is safe!");
+  }
+});
+
     }
     // ....................................................................
     // edit service
 
     const handelEditService =(id)=>{
+        // setShow(true);
+
         axios.put(`http://localhost:3030/services/${id}/update`,
         {nameOfService:name,image:IMage ,description:Description})
      .then((res)=>{
@@ -79,6 +99,9 @@ const handelAddNewService =()=>{
       }) 
 
     }
+
+    const handleShow = () => setShow(true);
+    const handleClose = () => setShow(false);
     // ................
     if(loading){
     return(<Loading/>)
@@ -159,35 +182,76 @@ onClick={(e)=>handelAddNewService(e)}
           <Card.Text className="texterea">{item.description}</Card.Text>
           <div  > 
        <Button variant="outline-warning" className="Cardbtn-1" 
-        onClick={()=>handelEditService(item._id)}>Updae</Button>{' '}
+        // onClick={()=>handelEditService(item._id)}
+        onClick={handleShow}
+        >Updae</Button>
+
+        {/*  */}
+        <Modal show={show} onHide={handleClose}>
+        <Modal.Header closeButton>
+        <Modal.Title>Update Service</Modal.Title>
+        </Modal.Header>
+
+        <Modal.Body>
+        <Form.Floating className="mb-3">
+          {/* name of service */}
+        <Form.Control className="Input"
+       id="floatingInputCustom"
+       type="text"
+       placeholder='Enter name of service'
+       onChange={ e=>setName(e.target.value)}
+     />
+     <label htmlFor="floatingInputCustom">Enter name of service</label>
+   
+   </Form.Floating>
+     {/* image of service */}
+     <Form.Floating className="mb-3">
+     
+     <Form.Control className="Input"
+       id="floatingInput1Custom"
+       type="email"
+       placeholder='Enter image of service'
+       onChange={ e=>setImage(e.target.value)}
+     />
+     <label htmlFor="floatingInput1Custom">Enter image of service</label>
+   
+   </Form.Floating>
+{/* description */}
+
+<Form.Floating controlId="floatingTextarea2" label="Comments">
+    <Form.Control className="Input"
+      as="textarea"
+      placeholder="Add Description"
+      onChange={ e=>setDescription(e.target.value)}
+      style={{ height: '100px' }}
+    />
+<label htmlFor="floatingTextarea2">Enter Descrption</label>
+
+  </Form.Floating>
+          
+          </Modal.Body>
+      
+        <Modal.Footer>
+          <Button variant="secondary" onClick={handleClose}>
+            Close
+          </Button>
+          <Button variant="primary" onClick={()=>handelEditService(item._id)}>
+           update
+          </Button>
+        </Modal.Footer>
+      </Modal>
+        {/*  */}
        
        <Button variant="outline-danger" className="Cardbtn-1" 
-      onClick={()=>handelDeleteService(item._id)}>Delete</Button>{' '}
-</div>
-        </Card.Body>
-      </Card>
-        
-        
-        
-        
-//         <Card key={index} className="box">
-       
-//        <Image src={item.image} alt='' width={200} roundedCircle/>
-//        <Card.Title className="title">{item.nameOfService} </Card.Title>
-//        <Card.Text className="texterea">{item.description}</Card.Text>
-// <br></br>
+        onClick={()=>handelDeleteService(item._id)}>Delete</Button>{' '}
+        </div>
+                </Card.Body>
+              </Card>
+          })}
+        </Card>
 
-//        <div  > 
-//        <Button variant="outline-warning" className="Cardbtn-1" 
-//         onClick={()=>handelEditService(item._id)}>Updae</Button>{' '}
-       
-//        <Button variant="outline-danger" className="Cardbtn-1" 
-//       onClick={()=>handelDeleteService(item._id)}>Delete</Button>{' '}
-// </div>
 
-//         </Card>
-    })}
-</Card>
+ 
 
     </> );
 }
